pyproject.toml
[tool.poetry]
name = "robusta-krr"
version = "1.6.0-dev"
description = "Robusta's Resource Recommendation engine for Kubernetes"
authors = ["Pavel Zhukov <33721692+LeaveMyYard@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "robusta_krr" }]

[tool.black]
line-length = 120
target-version = ['py39']

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
plugins = "numpy.typing.mypy_plugin,pydantic.mypy"

[tool.poetry.scripts]
krr = "robusta_krr.main:run"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
typer = { extras = ["all"], version = "^0.7.0" }
pydantic = "1.10.7"
kubernetes = "^26.1.0"
prometheus-api-client = "^0.5.3"
numpy = "^1.24.2"
alive-progress = "^3.1.2"
prometrix = "^0.1.10"
slack-sdk = "^3.21.3"
aiostream = "^0.4.5"
ipykernel = "^6.27.1"
ruamel-yaml = "^0.18.5"
pygithub = "^2.1.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.0.1"
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
types-pyyaml = "^6.0.12.8"
types-cachetools = "^5.3.0.4"
types-requests = "^2.28.11.15"
pyinstaller = "^5.9.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[project]
name = "robusta_krr"
